계산기. 파일분리
